Updated to give API switch
Created module folder for use with powershell gallery.

This includes a switch for setting/changing the API and has some checks to make sure the API is present before allowing the other switches to be used.

It also now shows details for asset tags that don't belong to the local machine regardless of the -Show switch being used or not.

It will also now check to make sure that the -Brand switch is only used when the powershell windows is elevated to avoid confusion.